module_1_traffic_detection/requirements.txt
ultralytics>=8.0.200
opencv-python>=4.8.0
numpy>=1.24.0
pydantic>=1.10.0,<2.0.0
pyyaml>=6.0.0
requests>=2.31.0
rich>=13.5.0
pytest>=7.4.0
